AdaptiveNormalizer
Added movingStats.h to .cpp file
