+++ src/AdaptiveNormalizer.h
@@ -1,0 +1,79 @@
+/*
+ * AdaptiveNormalizer.h
+ *
+ * (c) 2015 Sofian Audry        :: info(@)sofianaudry(.)com
+ * (c) 2015 Thomas O Fredericks :: tof(@)t-o-f(.)info
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef ADAPTIVE_NORMALIZER_H_
+#define ADAPTIVE_NORMALIZER_H_
+
+#include "PqCore.h"
+
+/**
+ * Adaptive normalizer: normalizes values on-the-run using exponential moving
+ * averages over mean and standard deviation.
+ */
+class AdaptiveNormalizer : public PqPutter, public MovingStats {
+public:
+  /**
+   * Default constructor. Will renormalize data around a mean of 0 and a standard
+   * deviation of 1.
+   * @param smoothFactor a parameter in [0, 1] representing the importance of new values as opposed to old values (ie. lower smoothing factor means *more* smoothing)
+   */
+   AdaptiveNormalizer(float smoothFactor=0.001f);
+
+  /**
+   * Constructor.
+   * @param mean the target mean
+   * @param stddev the target standard deviation
+   * @param smoothFactor a parameter in [0, 1] representing the importance of new values as opposed to old values (ie. lower smoothing factor means *more* smoothing)
+   */
+  AdaptiveNormalizer(float mean, float stddev, float smoothFactor=0.001f);
+  virtual ~AdaptiveNormalizer() {}
+
+  /**
+   * Sets target mean of normalized values.
+   * @param mean the target mean
+   */
+  AdaptiveNormalizer& targetMean(float mean) { _mean = mean; return *this; }
+
+  /**
+   * Sets target standard deviation of normalized values.
+   * @param stddev the target standard deviation
+   */
+  AdaptiveNormalizer& targetStdDev(float stddev) { _stddev = stddev; return *this; }
+
+  /**
+   * Pushes value into the unit.
+   * @param value the value sent to the unit
+   * @return the new value of the unit
+   */
+  virtual float put(float value);
+
+  /// Returns normalized value.
+  virtual float get() { return _value; }
+
+protected:
+  // Current value (normalized).
+  float _value;
+
+  // Target normalization parameters.
+  float _mean;
+  float _stddev;
+};
+
+#endif
