+++ src/PqExtra.cpp
@@ -20,26 +20,6 @@
 
 #include "PqExtra.h"
 
-AdaptiveNormalizer::AdaptiveNormalizer(float smoothFactor)
-  : PqPutter(),
-    MovingStats(smoothFactor),
-    _value(0.5f),
-    _mean(0.5f),
-    _stddev(0.25f)
-{}
-
-AdaptiveNormalizer::AdaptiveNormalizer(float mean, float stddev, float smoothFactor)
-	: PqPutter(),
-    MovingStats(smoothFactor),
-    _value(mean),
-    _mean(mean),
-    _stddev(abs(stddev))
-{}
-
-float AdaptiveNormalizer::put(float value) {
-  return (_value = MovingStats::update(value) * _stddev + _mean);
-}
-
 Normalizer::Normalizer()
   : PqPutter(),
     SimpleStats(),
